[roll] Roll fuchsia [kernel][phys] List supported test environments in qemu_boot_formats
This makes the qemu_boot_formats list element's provide a
supported_environments list.  The qemu_test_kernel() tests and
the shim tests now use this to plumb the environments list for
all the boot tests using each boot format.

This results in the phys tests for linuxboot and linuxbiboot
being run on AEMU as well as QEMU, and starts running shim tests
for linuxboot and linuxbiboot that weren't being run before.
When a crosvm_env becomes available for tests to run on, this
will trivially extend to having linuxboot64 and linuxbiboot phys
and shim tests also run on crosvm.

Original-Bug: 335547883
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049875
Original-Revision: dc2126a376f359da4019072ec24d4e1093689b9b
GitOrigin-RevId: 697dc01efc0ab431fc4dfe726b7188f97bf587ec
Change-Id: Ibac42dbc56e38750b12d2e175e32c1b3ff6eaeef