[roll] Roll fuchsia [gpt] Separate GPT to new rust crate
This is mainly a refactoring/reorganization to support testing GPT in
isolation, although as an added benefit the new GPT rust crate is now
portable if desired.

The rough split is that the GPT crate has business logic related to
parsing GPT metadata (and is set up in the future for having logic
related to manipulating a GPT as well), and storage-host just wraps that
and provides the "runtime" of exporting Block instances for the
partitions in a GPT device.

Original-Bug: 339491886
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105441
Original-Revision: 62f256a63431f6c703e3ee93453086fc3fd474ab
GitOrigin-RevId: 1255aca120347599e0309add6dcc9df432b54f59
Change-Id: I6a0fab54f47455730ecf91a14117e056125b169d