[roll] Roll fuchsia [ffx][build] Generate build-dir level config file
This generates the ffx build-dir level config file
as part of the build vs. being part of an fx command to run externally.
This makes it so the configuration will be the same for the same build
independent of the user's behavior or habits.

Original-Bug: b/360197942
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101656
Original-Revision: dbd7aa2033b26759c599f57cac61c1dd67746e58
GitOrigin-RevId: 956b2fd7dfbef9728251536e0664ef548352cf26
Change-Id: I00ea96ee4d8fa931052bf4f4a9fc8b7240589edc