[roll] Roll fuchsia [kernel][vm] Supply zeroes above content size
Adds an option for writing to a VMO to assume anything above the user
provided content size is zeroes, avoiding the need to read request
those pages. This is used by stream dispatcher when expanding the
content size on a write or append, as that range is logically zero.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1104694
Original-Revision: 105a389eeca6b72f6732a26dd66c864fbc9a4a92
GitOrigin-RevId: a4839af3a687b3d81723143dff7b8f1f807e70b8
Change-Id: Ifec7ac03a063c733eba443763b467aa3ed5d98a8