[roll] Roll fuchsia [netstack] Ensure buffer size fairness in benches
Inform both the TCP receive buffer size and the UDP send buffer size
when running socket benchmarks so that default stack values for those
sizes are not carrying disproportionate weight in the benchmark results.

While doing this, realized that netstack3 was not properly applying the
buffer size limits in all states, so introduced a new path for that with
a new test.

Removed the change detector from the tun socket benchmarks for the
socket sizes, it's not load bearing for correctness there and we can
get rid of some complexity.

Original-Bug: 42085918
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100455
Original-Revision: a20b1498cdc03c6f37312d87e00a9ba391dabaf7
GitOrigin-RevId: b24c01e18dc1d68942964c550808c56a4dd50f10
Change-Id: I8c95765143c374619cef45f6aeeddb079b9019b6