[roll] Roll infra/recipes [crosvm] Bump 'recipe salt'
Since we made fxrev.dev/1105592 without changing either the git
revision or the recipe salt, we did not end up pushing a new version of
the package to CIPD (since both it and the previous upload were tagged
with the same version ID). To fix this, we bump the recipe salt.

Original-Bug: 336934376
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1106592
Original-Revision: 114832b391f1369a4589c78c81ff05f4b92cbe63
GitOrigin-RevId: 0b797e192e602277c66c421da9a123a0a176c695
Change-Id: I917dfb9229ac166c8243a5bbe41f926d1e5af854