[roll] Roll fuchsia [storage][config] Use config capabilities
This migrates the majority of remaining uses for fetching kernel boot
args to config capabilities.

All new configs have their defaults described in the component manifest
and assembly routes a void when an option is not explicitly chosen by
via the platform config. This reduces the number of redundant places we
must declare the defaults.

The blobfs.cache-eviction-policy was removed from being set explicitly
as it defaults to NEVER_EVICT, which was the same option being set
explictly.

Original-Bug: 359305876
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105153
Original-Revision: 449ba2a5195b31cdf63dff51a5aa40c3771d4917
GitOrigin-RevId: 01af4b8111e0902749d5b9819d2ac255fd0c4dc8
Change-Id: I320a80c68c75c3a1aa5fc5a6759f0633f41b395c