[roll] Roll fuchsia [fuchsia.io] Add Open3 at all supported API levels
Add Open3 to fuchsia.io at all previous supported API levels. We cannot
safely call Open3 on components built against different API levels
otherwise. This should be the last time we are required to rewrite
history as part of the ongoing io2 migration.

To reduce the scope of this change, we just provide stubs in the
VFS that return NOT_SUPPORTED so that we can update the VFS libraries
separately. This will ensure that when communicating with fuchsia.io
servers, calling Open3 does not result in the channel being closed
due to an unknown interaction. This allows us to safely fallback to
Open1 by keeping the channel in a usable state.

Most partners integrating the SDK periodically update to new versions,
and the majority of fuchsia.io servers are backed by VFS libraries.
This means that once this CL lands and all pre-built binaries are
updated to use the latest SDK, we can begin calling Open3 safely.

Original-Bug: 348698584
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1102395
Original-Revision: 2943933ee8f4be914013d320f0dc4c675624f62d
GitOrigin-RevId: f938227ac0731422a4aac13a0d8bdfe223f31207
Change-Id: I0ee19412e393fbc685ec94226294406e56b9f092