[roll] Roll fuchsia [amlogic-display] Post asynchronous tasks to set Vsync receiving state.
Previously, when disabling Vsync receiving, the Vsync receiver
synchronously waited on the Vsync loop dispatcher to make sure
the Vsync wait is cancelled. This caused a race between the
coordinator Controller dispatcher (which calls SetDisplayPower())
and the Vsync IRQ handler dispatcher:

On the Controller dispather, SetDisplayPower() acquired the
display_mutex and then synchronously waited on the Vsync receiver
to finish cancelling IRQ waits.

On the Vsync IRQ dispatcher, the IRQ handler was waiting for the
display_mutex, and the async task to cancel the IRQ waits was yet
to run on the dispatcher.

This caused a deadlock on both dispatchers.

This change fixes the issue by making SetDisplayPower() non-blocking.
It only posts an async task to begin / cancel the IRQ wait but not
wait on the async task to finish.

Test: On core.vim3 with AML028-30MB-A1 HDMI display,
 1. Run "display-tweak panel --power off", display is turned off
 2. Run "display-tweak panel --power on", display is turned on with
    previous displayed contents.
 3. Repeat steps 1 and 2 multiple times.

 Testers can add a zx::nanosleep for 1 second at DisplayEngine::
 DisplayEngineSetDisplayPower() (L738) to make sure a Vsync event
 is handled during SetDisplayPower() and it doesn't cause a deadlock.

Original-Fixed: 361196365
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107983
Original-Revision: 2269f1e2c37c183ab3b1f5dcde30a7d102e4bbc2
GitOrigin-RevId: cca92c1f2b2c9b7d402c5aa3415f651305735d23
Change-Id: Ie93d1f1d80a841eb0ed04f711544764629eb0dc0