[roll] Roll fuchsia [aml-cpu] Return an error if voltage could not be scaled.
The CPU driver calls out to the Power Driver to set the voltage.

Currently there is a bug that if the power driver returns a voltage
different than the voltage that was requested, the CPU driver simply
returns ZX_OK. It currently may not even scale the frequency and
just silently return success without accurately setting the operating
point.

There is another bug that if the frequency is lowered but the call to
lower the voltage fails, the frequency is not restored to the original
value.

I added two new test cases to aml-cpu-test.cc that demonstrate the
failure mode and then fixed aml-cpu.cc such that they pass.

Original-Bug: b/309798380
Test: aml-cpu-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108079
Original-Revision: bcd1715ae48029de942e127a3a02f6079c30ee45
GitOrigin-RevId: be3501ab509fef41f234561ca7aecb531aaf1367
Change-Id: I17f545e27d48f21223a439f886aeb0866f34c28d