[roll] Roll llvm-project/libc [libc] Fix file collision causing test flake (#106119)
In patch #105293 tests for vfscanf were added, meant to be identical to
the fscanf tests. Unfortunately, the author forgot to rename the target
file causing an occasional test flake where one test writes to the file
while the other is trying to read it. This patch fixes the issue by
renaming the target test file for the vfscanf test.

GitOrigin-RevId: 32e6827b71850c5508a455fe49812fb562ce1d92
Original-Revision: 4f66756562a69c236872acd2a73c1dff2417a278
Change-Id: I1b8243f5b78b29bfd682314da20c0629c1ffebb3