[roll] Roll fuchsia [netstack3] Remove Ethernet filtering
Remove support for filtering at the Ethernet layer. This better conforms
to Linux & Netstack2 behavior, which do not perform L2 Filtering
directly within the netstack. Rather, they rely on device drivers to
perform L2 Filtering.

In the absence of L2 filtering support in the device driver (currently
the case for Fuchsia) the packets will still be filtered appropriately
in the IP layer.

The motivation for this change is multicast forwarding: There are
multicast forwarding integration tests that rely on the Netstack
delivering multicast packets to the IP layer, despite not having
link-level multicast interest in the packet.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109272
Original-Revision: cacd7dd4a514fd6e217963c85fb2f6725d759c63
GitOrigin-RevId: 096400ff8f58f8a9aea39e576f4c5e7898a1aa57
Change-Id: Ieb91d944b52a6be665bbfc9243984404ebf377dc