[roll] Roll fuchsia [fidl][cf] ComponentController open
ComponentController adds a couple of events (OnEscrow, OnStopInfo) at
HEAD. To make the NEXT API level ABI compatible with the PLATFORM API
level the ComponentController protocol needs to be ajar or open and
the new events need to be flexible.

This CL updates the FIDL declarations to make ComponentController open
at NEXT and the OnEscrow and OnStopInfo events flexible. It updates
various clients and client libraries to reflect this.

While we could have just made ComponentController ajar, we think that
we'll want it to be open eventually, so this goes the whole way.

Original-Bug: 344640629

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105374
Owners-Override: Adam Barth <abarth@google.com>
Original-Revision: fb10e45db218af5e9203506efa85722976d06c63
GitOrigin-RevId: 2c82630803384ac4ee6b063f28a2417e24ac566c
Change-Id: If0389b98453ede4c2fdd5c14e5ef78e1bbd15c33