[roll] Roll fuchsia [bt][hfp] Send AudioControl connect/disconnect
When a peer connects, an AudioControl should receive a call to
connect to possibly set up audio system indicators that the peer is available
for HFP usage.  This enables the audio system to start/stop the HFP
audio without a call in progress.

Also send a disconnect when a peer is disconnected, to tear down any
active audio connection and/or make it unavailable

Original-Bug: 341116731
Test: added new tests, fx test bt-hfp-audio-gateway-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065732
Original-Revision: 62d15a98707979a2f8fc3f217931aba12eb750d3
GitOrigin-RevId: 92501d26a8609642f08990617f074810b7f7effa
Change-Id: I21e9960fe3cbe3767de80db95c40223c7e0a06e6