[roll] Roll fuchsia [syslog] Remove redundant thread koid TLS slot
We already have a thread-local storage slot for the thread koid in
logging_backend_fuchsia_globals.cc. That slot is linked into a share
library, which means it is shared for the whole process. The slot
removed in this CL is in logging_backend_fuchsia.cc, which is in a
source_set. That means this slot is duplicated in each linkage unit that
uses the source set.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109348
Original-Revision: 211b5d4c38e9b5448504190b1d3bf2e6eb02ebfc
GitOrigin-RevId: 3eda47299ba3f416825a6739516500eb472498cb
Change-Id: I1fc14faa009d72cd716a7e2b50ed63b320a91e58