[roll] Roll fuchsia [trivial-allocator] Clarify PageAllocator<Memory>::Allocation release semantics
Previously, the releasing in
PageAllocator<Memory>::Allocation::release() was accomplished by calling
the destructor of the held Memory::Capability. We anticipate a use-case
where it's natural to have Memory::Capability as the phys Allocation
class, but that frees - not releases - on destruction.

In this change we update the expected Memory interface to include a
Release() trait, and then use that in release() above to give clearer,
release semantics and a more flexible layer of indirection that enables
the use of using Allocation for a Capability.

Original-Bug:
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108936
Original-Revision: 8a5826c985764be5c05ae3df8e32c848e865eb44
GitOrigin-RevId: 427ff425dc52906c62e5517aaaf3992f517ef905
Change-Id: Ib019dbb6c8870d920addd84bc5e37e2fe5f16268