[roll] Roll fuchsia [fidl][audio] open RingBuffer, signalprocessing.Reader & composers
The fuchsia.hardware.audio.signalprocessing/Reader.WatchTopology() and
fuchsia.hardware.audio/RingBuffer.WatchDelayInfo() methods are added at
HEAD. To allow platform builds to remain ABI compatible with SDK builds
these methods need to be flexible and so the protocols they're in need
to be open. Also all of the protocols that compose those open protocols
also need to be open.

This makes the changes at NEXT so that the next stable API level will be
ABI compatible, and updates the in-tree users of the protocols to work
with open protocols and flexible methods.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105376
Original-Revision: 9f00b80ecce906b6f9b96002e76940326e493c9b
GitOrigin-RevId: 6de36ee6cfec6dbc8a6411941881e2db9f7362fa
Change-Id: I0cca5b6eb2ae940c0f26d1c9bf13a63b7b31cd82