[roll] Roll fuchsia [kernel][phys] Stop bloating the physboot-elf-kernel-test kernel package
Previously, physload would reuse the space previously occupied by the
kernel package (i.e., the STORAGE_KERNEL ZBI item) for the hand-off to
the kernel. With profile instrumentation, this hand-off further encodes
profiles which can take up a fair amount of space, and as a result we
were running out of space for it in the case of the
physboot-elf-kernel-test (which naturally had a much smaller kernel
package than the main Zircon one). We mitigated this by artifcially
inflating the kernel package by sticking Zircon in there. However, now
that hand-off is no longer allocated in a limited way, we can undo this
hack.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100000
Original-Revision: 8b006dcd74405d26486dc2eee2b98c5de7d17ab2
GitOrigin-RevId: 396a0e88c0e34bc1560c6fddd20035bdd1f79674
Change-Id: Id135f03e5bd07ec561b9758400cca30c6c83de96