[roll] Roll fuchsia [starnix][iptables] Translator creates rules with reasonable indices
fnet_filter_ext::Rule requires an index field that has a unique u32
w.r.t. other rules in the same Routine, but does not need to be unique
w.r.t. all other rules in the same Namespace. Previously, we used the
the entry_index of the Entry that the rule was translated from. This was
problematic for tests because it exposes the internals of the
translation layer, and depending on the table, adding a new rule on an
otherwise empty table resulted in different indices (e.g. 5 for mangle,
3 for filter).

Additionally, fnet_filter_ext::Routines created from the translator is
now ordered. Installed routines are defined first, in the order of
Prerouting, Input, Forward, Output, Postrouting, if they exist on the
table. Custom routines are defined after, in the same order that
iptables receives them.

To simplify logic, `RuleSpec` is removed, and `Entry`s that represent
rule-specifications are now grouped by the Routine that they are part
of - both `InstalledRoutine` and the new `CustomRoutine` own a
`Vec<Entry>` for the rules they own.

- Removed entry_index from `Entry`, as it is no longer needed.
- Removed routine_map from `IpTable` translator. Previously, routine_map
  stored both installed and custom routines for JUMP lookups. This was
  unnecessary as Action::Jump only requires the name of the Routine, and
  installed routines cannot be JUMPed to. Replaced with simpler lookup
  table.
- Removed `RuleSpec` and `UntranslatedRule`. Unsupported rules are now
  stored as a None-variant Option, and are filtered out before rule
  generation.
- Added `CustomRoutine`, similar to `InstalledRoutine`, now own the
  entries that belong to them. This makes the code easier to reason
  about.

Original-Bug: 354763810
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108435
Original-Revision: b10c2c91ffd30ae0298ce29cd1575b81c34a4a55
GitOrigin-RevId: e1e8a000912aa1b754f4600956e2d0367775f8d5
Change-Id: I8deab5e5d4b40597ac344f988536203ddd5dcad9