[roll] Roll fuchsia [fuchsia-async] Change scopes to join when task finishes
This CL changes it so that scopes are joined when the task finishes
rather than when they finish *and* their handles are polled.  Waiting
for their task handles to be polled shouldn't be necessary and might be
confusing to users.

This also changes `Task::cancel` so that it doesn't need to be polled
for it to actually cancel the task; callers can ignore the returned
future if they like.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108273
Original-Revision: 6257a5eff9c994deb5d19102eecf5814415f2423
GitOrigin-RevId: 69fb80296223c11aa04d239dbe753909b183f57a
Change-Id: I0aa148c07b19a74da9c44a1413b1d1299d5ddacd