[roll] Roll fuchsia [driver_manager] Update test fixture's StartDriver impl.
CallComponentStart has been moved to the fixture file
and renamed to DriverHostComponentStart.

In the case of a non-colocated driver using dynamic linking,
we want to call the driver host runner Start before returning
from the StartDriver function. Otherwise the driver won't
yet be started (waiting for the driver host to start),
and the returned started TestDriver will not be a valid
pointer.

Original-Bug: 355233670
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108572
Original-Revision: c5936d20778204260eefa183ff0076cec38e16a2
GitOrigin-RevId: b718564c4dd97f006e889fe058df4b69bdf731ec
Change-Id: I50ffa25256311e534c540f8d7efc2543ec4828e8