[roll] Roll fuchsia [debugger] Fix crash with unverified line table entries
During the search for the best match for a file+line based breakpoint,
it was possible for the line table to match an address that corresponds
to no code, which produces an invalid Function symbol. The typical case
for this happening is an address that corresponds to some compiler
generated code.

Original-Fixed: https://fxbug.dev/352164877
Original-Fixed: https://fxbug.dev/360821506
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110912
Original-Revision: bda367e7a8abc7226ee75dce98de633936e0eda6
GitOrigin-RevId: 4ebfe388d50015d2a82e2b36e79ba953cc4bfb46
Change-Id: If7cb7fe93ccbfe69cf509697cb0965936bc7b9f6