[roll] Roll fuchsia [ffx][repo] Add autopublish option to server start
This adds an option to server start, that takes a package list
file  be watched and automatically publish the packages as they change.

This is the same implementation as `ffx repository publish --watch`,
just conveniently part of the repo server start command.

It is primarily used in-tree, to simplify the implementation of the bash
wrapper script `fx serve` and to keep the publishing log messages
collected in the same file and streams as the repository server log
messages.

Original-Bug: b/360933107
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109337
Original-Revision: fd935282634dd3e8a7cc22bb92e5ab403d023c31
GitOrigin-RevId: f5f4b4a02ff21435e2bc34fe5b03e777b3a0d630
Change-Id: Ia836c88b906019cbc5512d8bc01b2ae548756a7c