[roll] Roll fuchsia [wlan] Fix components double remove test flake
One of the tests in the wlan drivers component library was intended to
test that calling Remove twice in a row would correctly handle the case
of another removal being in progress. Unfortunately this test made some
incorrect assumptions about the internal state of the NetworkDevice
object it was removing. The test assumed that the internal state was not
updated until after the removal notification. This is not the case and
this lead to flakiness when the first removal completed before the
second Remove was called.

Instead of blocking the removal notification, block the entire
dispatcher used for removal with a posted task. This properly prevents
the first removal from completing until a completion is signaled.

Verified by adding a sleep between the removals which always triggered
the flake before the fix and passes after the fix.

Original-Fixed: b/338922519
Multiply: wlan-drivers-components-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1111818
Original-Revision: 84723ec6762916199564e3a7e8de7435d1ddc64a
GitOrigin-RevId: 4633d146b6a91dd0a65e56f66ddd96008b67d273
Change-Id: Ia62fc2d166b74caad6c71449998baea98ab2083a