[roll] Roll fuchsia [windowed-stats] Introduce a time series glossary.
This change adds a glossary that describes the concepts and components
of the `series` module. These ideas are not trivial and this provides a
much needed reference.

This is more or less a first draft; expect subsequent changes. It is a
living document that should evolve with `windowed-stats`.

Original-Bug: 350797382
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108023
Original-Revision: cf64f18687c1cd92878abfce29c6b0d28ca5e29b
GitOrigin-RevId: 09eb51e6c8282692df3144b4e6345430979eeb9d
Change-Id: Iac247b2cfea65f6214051d5d3df167947ea89f1c