[roll] Roll fuchsia [bazel] Generate @fuchsia_idk repository.
Add a new repository to the Bazel workspace used by the
platform build, to expose the platform IDK as an external
repository (currently named @fuchsia_idk).

This still takes a Ninja-generated IDK export directory
as input, but will rewrite the metadata JSON files to
include labels, instead of relative file paths, when
addressing Ninja-generated artifacts (with a few
exceptions, such as a few files in the sysroot).

A future CL will adjust the implementation of the
fuchsia_sdk_repository() rule to accept this external
repository as a proper input.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110273
Original-Revision: 2c9c7590ff10f12384cf2bef24f6915a3a5dd291
GitOrigin-RevId: b382e92cc689ec8e24e19215905b76466ea3e45a
Change-Id: I0992d19a19b6c5d8c454a167532e55a546b3dd0b