[roll] Roll fuchsia [inspect][graph] Fix issue that was leaving TOMBSTONES behind
There were multiple issues surrounding the RAII semantics of the
Vertices and Edges. In particular, dropping either an edge, a source
vertex or a destination vertex was resulting in the edge being partially
dropped.

This change removes the usage of clone_weak which was the source of
problems here. We may want to entirely work on removing that API or
providing better support for it in the inspect library. It's very easy
to misuse in subtle ways.

Original-Fixed: 363073059
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112732
Original-Revision: 8f5297a595f28d9ebe4f1a542693955ffef76eb1
GitOrigin-RevId: 41d2d3bfee7584f3e04d648659ace9307f208a91
Change-Id: I5d57a8627118148475fa2a00174a074f24852273