[roll] Roll fuchsia [kernel][rppm] Introduce PowerState
PowerState is a member of a device(cpu) that belongs to a power domain.

The PowerState sole duty is preserving a coherent state of a meaningful
power level.

A power state is represented as a triplet between <Model, Actual Power
State, Desired Power State>. When a model is updated to a functionally
different model, then both actual and desired power level lose meaning.
It is the job of the PowerState object to determine whether a transition
should be dropped or not based on whether it is still relevant for the
current Model.

Additionally a PowerState can request a `Transition` which generates a
`PendingTransition` encapsulating all the information needed to perform
a transition at a later stage.

Original-Bug:
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089296
Original-Revision: 10730064da1b2693219c6a7562fcf15408ceb590
GitOrigin-RevId: ce9efef0ead1d88385f80da3e2a737ba0465dfd0
Change-Id: Ie6ccc681595730d3759d9cae2f1f73edbab0bb9e