[roll] Roll fuchsia [netstack3] Move `confirm_reachable` to `IpSocketHandler`
We need to lookup route to confirm neighbor reachability because we
don't have route caching. Moving it to the IP socket layer will make
it closer to where routing decisions are made, and in the future when
we have route caching, it will also be in the `IpSock`.

Original-Fixed: 361817008
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109983
Original-Revision: f38ea34b6495c054e12600939983129cf2b2e67c
GitOrigin-RevId: be4bb51096896bb107cf8f0396dbc4dd7ddff48d
Change-Id: Id4142a42aa31f26bdf20f4e9cc34b277b675463a