[roll] Roll fuchsia [fx] Fix build-dir level default targets
This aligns the fx scripts with how ffx works so build-dir level
targets can be configured, but are still overridden using
`ffx target default set` or the `--target` flag.

This includes the -d flag to `fx`

ffx uses a prioritized list of configuration to resolve the default target.
Once there is a value, the remainder of the list is ignored.

* Specify the target on the ffx command line with --target.
* If configured, use the user level configuration set by ffx target default set.
  This user level configuration is stored relative to $HOME.
* The default configuration compiled into ffx checks two environment variables:
     "$FUCHSIA_DEVICE_ADDR"
     "$FUCHSIA_NODENAME"
    These are used by Fuchsia infra jobs to specify which target should be used
    when running a specific test FUCHSIA_NODENAME is also set to any
    build-directory default target set by `fx set-device`.
* If only one device is discovered

If there is still no target device identified, or if there are multiple devices discovered, an error is returned.

Original-Bug: b/362754361
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110007
Original-Revision: b392d7a90bdd0a9a948fe478e9e25c9ab0683798
GitOrigin-RevId: eee008b2687f3d621d559974fd35ebf8a55dacca
Change-Id: I6e36e29637da5a17c6a43954183ec2090d4a7df2