[roll] Roll third_party/Vulkan-Loader [clang][VulkanLoader] Suppress warnings from -Wunused-command-line-argument
After https://github.com/llvm/llvm-project/pull/101103, clang will no
longer accept mixing `-S` and `-c`. An upstream discussion on shows that
this was intentional, and is unlikely to be reverted:
https://github.com/llvm/llvm-project/issues/104347. It also does not
appear to be possible to remove the `-c` option from the compiler
command. As a compromise, we opt out of the option warning so that
-Werror will continue to work.

Original-Fixed: 359934302
Original-Bug: 360223478
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/Vulkan-Loader/+/1101677
Original-Revision: 710b06e7ff8e672b7359489fe4906f705c2761d4
GitOrigin-RevId: fbf6944d66d8a02abda9b9be33273474892cf8ca
Change-Id: Ib5b0e17ccf55c9746f6b9b126a93bbfbe8ab0f65