[roll] Roll infra/recipes [fuchsia/build] Only set one multiplier comment per patchset.
Although multiplier comments can be useful to notify that the multiply
footer actually worked and show all the builders where it ran, it can
get excessive and noisy if there are a lot of builders that run the test
since it adds a comment and thus email notifications for each builder.
Instead it will probably be enough to only show one instance of a
builder that ran the multiplied test on a specific patchset.

Original-Bug: 364353119
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1113458
Original-Revision: 756fa08afe70d84c6b9a3c028c108e37f467c0da
GitOrigin-RevId: 6cf186f419999d16dfdf7145f4936204f15fbe80
Change-Id: If0dcce9e29b5c3b9a27c07d7d37a4ec5d9d43dc9