[roll] Roll fuchsia [ufs] Support UTP Task Management Request
This CL defines a UTP Task Magenemt Request Descriptor and creates a
TaskManagementRequestProcessor class to manage the list and handle the
doorbell and completion.
The Task Management Request is not yet issued to the UFS Device and
will be used in the error handler that will be implemented in the SCSI
libray.

Test: fx test ufs-tests qemu_ufs_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992272
Original-Revision: 826af6b9e3822fd4bd5fa549de393f51ae7d3068
GitOrigin-RevId: 71c21da0f7327db0abac7b56b5eb680f698bf16d
Change-Id: Ibf9487734dcb2da608b4e049f64ab66b24a72f57