[roll] Roll fuchsia [testing] Add a template for c++ test executables
To enable LTO in the platform, we want to be able to disable LTO
during the link step for test executables, in order to save
build time (this requires the use of FatLTO).

This new template declares that an executable is a test
executable, and should receive different build optimizations
than a standard executable receives.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105436
Original-Revision: 968bbd4a9540468c5a606410d97be30d0cb82f37
GitOrigin-RevId: 147ecd523eb9ddc9128fb35110874d1e3d1e0e05
Change-Id: Ic7738524bb43dbc139771b99a8a7764211d6f1f6