[roll] Roll fuchsia [starnix] Add wake locks to the kernel
This allows the kernel to keep track of active wake locks and/or active
EPOLLWAKEUPs.

An EventPair is signaled whenever a wake lock is active, and cleared
whenever the number of active locks drops to zero. This allows the
Starnix runner to inspect the event to determine whether or not a wake
lock was acquired during container suspension.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112632
Original-Revision: fe3ad2924c4db0fa227a4fe30a74772d60d95a3c
GitOrigin-RevId: 63b68cd4e8f20433b679c7835c3224a304c60cf1
Change-Id: Ied5eb125280549422d0433a6b01aab4101681366