[roll] Roll fuchsia [bt][gatt] Remove unnecessary handle tracking
Discoveriing a handle for the service changed characteristic is used to
decide whether to attempt to send indications, but is never cleared when
there are no subscribers, so will likely be on after the systems runs
for any reasonable length of time. Instead continue to use the set of
subscribed peers directly.

Test: code removal only; fx test bt-host-gatt-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009175
Original-Revision: 01c53083ebf40906508a6126e143651a8c5f94db
GitOrigin-RevId: 990212711df445c61486991a3119ed26b34cc039
Change-Id: Iae1c8c5669d4075a8e8f6697d99ce83f0dda29af