[roll] Roll fuchsia [build][api] Fix `print --last-build-only product_bundles`.
The command was always returning an empty list, even when building
an actual product bundle in the last build invocation, because
it filtered based on the "path" value of each entry, which points
to a directory that never appears as an output path in the Ninja
build plan.

To fix this, filter on the "json" field instead, which points to
the actual product bundle JSON manifest file, a real Ninja artifact.

Original-Bug: 365039385
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114412
Original-Revision: 91508af8c6bae19f6effa77ae6cbe07f0e403c07
GitOrigin-RevId: f8d702f54583e0225a65656606d75a40d3515ecf
Change-Id: I6e9e687d541ab0bafdfa10595c38b0973dfde19c