[roll] Roll fuchsia [wlan][policy] record if network is single BSS
This change records whether multiple BSS's have been seen for a network
so that roaming can avoid scanning when there are no other BSS's to
choose from. The implementation will wait for a few scans that have only
one BSS before deciding that the network only has one BSS, and it only
pays attention to the network that is being scanned for. The data is not
persisted.

Test: unit tests: fx test wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110344
Original-Revision: e5427b343b5f90ddb7fcb51453d6379161fa5296
GitOrigin-RevId: 5080c6344bac75ff42b4326db31658138c723c8e
Change-Id: I843f01fa013e8ce5f7c83b6cb601c36addde58d9