[roll] Roll fuchsia [driver_manager] Implement DynamicLinkerDriverHostComponent Start.
This makes a call to fuchsia_driver_loader::LoadDriver,
(currently returns ZX_ERR_NOT_SUPPORTED).

Adds a partial fake driver host, to be completed in a future CL.
This is so the bootstrap channel is not dropped prematurely,
leading to the node beginning removal and the LoadDriver FIDL
request not completing.

Original-Bug: 341998660

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093155
Original-Revision: c907229cb0b15dda715191a9d03c68306faaf828
GitOrigin-RevId: 238b0fa264f4136a978d0fa60c730a96f547acbc
Change-Id: I3818dc5767ff32f19039a0424513b808ac7abb5f