[roll] Roll fuchsia [fasttime] Use sequential read for monotonic ticks
The mono_ticks_offset in the vDSO is no longer constant, as it will be
updated during periods of suspension. Therefore, it must now be read
with a sequential read just like the kernel version of the same routine.

Note that the register read ordering is still not correctly enforced;
this will need to be fixed in future CLs.

Original-Bug: 341785588
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114705
Original-Revision: 889fa9d359d861fc7c49cb62e58dd0f9832fe927
GitOrigin-RevId: c42a13b362569a48ee4d029fa19c76b971cf3ddd
Change-Id: Ib587b36a8e22cce3be2542bae4eb13c96e9be452