[roll] Roll fuchsia [device-enumeration] Don't verify partitions
Which partitions get bound is more of a question of whether fshost is
doing its job properly, and we have extensive tests for fshost, so
remove these from device-enumeration.

This is necessary for enabling storage-host, since that feature will be
enabled on a per-product basis, not a per-board basis (and so we'd have
to change which devices we look for depending on which product is
running the test, which is tricky because the test is compiled the same
regardless of the product it is being compiled for).

Original-Bug: 339491886
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115814
Original-Revision: 6f074f8fdc99af835a95a3cfdd88093495c7ec6e
GitOrigin-RevId: feaa7dd0f47cace8c1a47665bdce2b6c616090c7
Change-Id: Iebdda930412973872b33ceb77dc75bba1d63d7dc