[roll] Roll fuchsia [bazel][assembly] Add missing "json" metadata field for product bundles.
The schema of the `product_bundles` build API module states that
the "json" attribute, which points to the product bundle manifest
file, is required.

However, the bazel_product_bundle() didn't define this in the
metadata it generated for Bazel-assembled product bundles. This
caused `build/api/client print --last-build-only product_bundles`
to fail, because it could not find the "json" object field when
trying to filter the content of `out/default/product_bundles.json`.

See the CQ failures on [1] for examples.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1114414/2
Original-Bug: 42175326

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116274
Original-Revision: 48085863a8d397052428e8062914a945f96d2a0e
GitOrigin-RevId: bc8c8d72fe0ab82ae998183d48ab369e4d29f9d1
Change-Id: Id102a790073aa42306bf4fb255aac6330febe812