[roll] Roll fuchsia [sestarnix] Use a track_permission_stub macro for unenforced checks
Introduce a macro akin to track_stub!(), for permission checks that
are not currently enforcing, to ensure that they are not forgotten
about.  Each permission check "stub" must start with a TODO() which
will be passed to track_stub!() to allow tracking of the number of
implemented-but-ignored checks.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114392
Original-Revision: abcf41ce3a5526b3d0eee1091ebd21631cc9a59f
GitOrigin-RevId: e176bd95b7767f8b06b617a1e39bf514eae37f14
Change-Id: I2fffef690fd066090ad6e7503918cf65d531e0ea