[roll] Roll fuchsia [starnix] Reflect reserved brk space in mappings
Starnix reserves some space after brk so that the addresses from that
range are not selected and brk can be safely grown.

In preparation for switching to Starnix-based address selection, change
the logic so that this reserved space is tracked by Starnix (and not by
Zircon only as it was before).

Additionally, this makes this behavior consistent across forks, as
previously we were not replicating the reservation.

Original-Bug: 309452853, 365834029, 333111647
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116297
Original-Revision: 7d3562e0981993fa6c5875f43fb033b56cdade91
GitOrigin-RevId: 4a8fc7e51d8345f9b2f4f024ee8351f1e70710e1
Change-Id: Ifef0abcef5b6467ad753447f0f30c2c1624a1cc8