[roll] Roll fuchsia [ffx] extend daemon connection timeout
We see numerous cases in Infra where a 5-second timeout for connecting
to the daemon is exceeded, leading to numerous flakes. Extend the
timeout to 15 seconds to avoid many of the current flakes.

(This also fixes a bug in Fuchsia Controller in which the config
system wasn't initialized.)

Fixes: 365631259
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116693
Original-Revision: 45268db7548f8742170564b143750e992b340164
GitOrigin-RevId: bc600cd03b81b7ca8ff9e10dd5865f2583f8eb0d
Change-Id: I9c3b2e7b3bdce437f6789c5f974dca8da5b5ce9b