[roll] Roll fuchsia [fidl] Rename fuchsia.device.manager to fuchsia.system.state
The old library was based on an older name for the fuchsia driver
manager. However these days the only protocol that remains in it is
served by the shutdown-shim component with clients that are not driver
manager. Renaming it to more accurately relfect what it is should help
reduce any confusion with respect to it's purpose.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115616
Original-Revision: 40fea725e4025042106c14e45f9b166617d6bb9e
GitOrigin-RevId: f91e703baf2ffa8181450504193b2978aea3f844
Change-Id: Idb48fc53df721454a5b1c089311f0da536593d19