[roll] Roll fuchsia [starnix] Implement ProxyWakeChannel in runner
This implements the new ProxyWakeChannel in the Starnix runner. The
runner creates a new thread which waits for either of the proxy
endpoints to become readable (or receive a peer closed). When a message
is available on one endpoint it is proxied to the other.

If a message is proxied into the container channel endpoint,
EVENT_SIGNALED is raised on the event provided to the ProxyWakeChannel
call. The will wake the container if it is suspended. Future suspend
attempts will fail until the client of the ProxyWakeChannel call clears
the signal from the event.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117712
Original-Revision: 6f9a51a35e6be003282f8e04e39e29dda8f5bfac
GitOrigin-RevId: c96ddfc12e8627d50c4ce05a751fe117e3963e1f
Change-Id: I93a58f641a43e1253c219a6177c61a4b20c78a65