[roll] Roll fuchsia [physload] Test harness will own the kernel storage.
The API for a physload handoff test consisted in handing off the kernel
storage to the test. This is problematic because, when the test routine
returns, the kernel storage is released (the allocation member is
destroyed). Then the test harness prints the routine, and follows up
with calling abort.

Why is this problematic? Well any interaction that may allocate memory
after the test completes, may allocate over the image of the currently
executing image.

This introduced a source of flakyness for this tests, because depending
on how long it took to dump the stack (in this case), it would fire a
tefmo check against failed assertiongs.

Original-Fixed: 366078334
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117532
Original-Revision: 9f6ac196ab42ffd5a794631e9012580c3c5bdbcc
GitOrigin-RevId: 8715b151e1cd1aaa4956c686f2409237a0f8ed97
Change-Id: Ie640d9b2d23911e2f831fa9361cac800f8baa526