[roll] Roll fuchsia [assembly] Give developer overrides allowlisting a better error message
The definition of developer overrides targets is currently restricted
to local-only locations (they cannot be checked into the tree and used
from there).

This is to surface use-cases and discuss whether they are best
defined as local settings, or features that need to be added to
the platform configuration schema and rules.

This change gives this restriction a better error message than
the 'dependency not allowed' visibility restriction that it
currently has.

Original-Bug: b/366083446

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117776
Original-Revision: 0a519a95f62f1b0876293ce08b6fec197fe33446
GitOrigin-RevId: 1a5b20a5db62954d8bd798393ee717ac777851b7
Change-Id: Ifca253ead0118c03ba850012a8b1865f9c9703df