[roll] Roll fuchsia [ffx][emu] Add context to 'persistent emulator'
Due to the ordering of the subcommands in the ffx documentation ("start"
comes before "stop"), at the point of reading through the `reuse`
switches it is unclear what a persistent editor means. This adds a key
piece of helpful context for the reader in the documentation of the
"start" subcommand.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116493
Original-Revision: c7638c43b0bb084ea9644f9054797d02a850da89
GitOrigin-RevId: 29654a4be058a3b09961fe27c1dc2b101dd70cfc
Change-Id: I792488c03fcff6a1d82a72a86f38e28c7dfa1a97