[roll] Roll fuchsia [sestarnix] Fix SeLinuxApi<> seek handling for reads & writes
The "selinuxfs" has a number of files implementing administrative
and query APIs, which are seekable, but have asymmetric behaviour,
with `read()` operations behaving as for a normal file, but
`write()` operations not updating the seek position, and varying
in whether they ignore or error if made from a non-zero seek
position.

Most fundamentally, this fixes the implementation to correctly
return end-of-file, rather than mistakenly returning the same
content over and over again.

Documentation of the expected behaviour of the SeLinuxApi<> helper
is updated to reflect the behaviours observed with the Linux
implementation.

Original-Bug: 365812498, 304775883
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116496
Original-Revision: 29d38711b227846a850635751e94d732cee52e36
GitOrigin-RevId: ddd915330d4662c953b56ffa9c16a80b739c9a97
Change-Id: I129929e40344230a63d89dc5773ee7271f69d1fd