[roll] Roll fuchsia [wlan][fidl] Make FIDL libraries versioned
Add `@available(added=HEAD)` to libraries in the "fuchsia" namespace
that are included in an SDK [category] but do not currently specify
`@available`. This will allow the `fidl()` template to verify all
libraries in the "fuchsia" namespace in a subsequent CL.

In one library, `7` is replaced with `HEAD`. The `7` was added in
https://fxrev.dev/664072 but not removed when the category was fixed in
https://fxrev.dev/679350.

Original-Bug: 364628320
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114043
Original-Revision: 2cb818e48a2c24234b0dfabcf98c3c1f66f7c857
GitOrigin-RevId: 5216c58bf3ed4e938c14a6201502e42e6c7a8d0f
Change-Id: If2af286e7018d4214ed75245d432d1fe866b08d4