[roll] Roll fuchsia [boot-shim] Do not set up MMU on riscv64
The shim does not need the MMU, and when the MMU was set up we have not
decoded the extension/feature set. In some cases this lead to setting up
invalid bits in the page tables causing an exception on boot.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115955
Original-Revision: c6c2c42e6f9d4de023cfd98fa4206cdb20e990ba
GitOrigin-RevId: 5079fc50cf277cf7834c6b282357e16a13c8098b
Change-Id: I7bb2b3b4641fc9151ab762caf7e0e4de1472f667