[roll] Roll fuchsia [dtr] Use test_package field instead of pkg
The new test_package field is now part of the partner sdk and supports
both the subpackage workflow as well as the old workflow where the
driver is in the test package directly.

This also adds a new field to the start args called
boot_driver_components which can be used to decide if a driver should be
a boot driver, aka started in the boot collection which has more
capabilities available to it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113381
Original-Revision: 23a60edb2d9e3be5d72314f5be9cc018bf881814
GitOrigin-RevId: 3746861c31c62b75c021411d2112fd89fad0abe4
Change-Id: If5066455fc547a1a4988dd32e4a56a8e7b9e8d78