[roll] Roll fuchsia [ui][input][fidl] Make all fuchsia.{input|ui}.* libraries versioned
Add `@available(added=HEAD)` to `fuchsia.input.*` and `fuchsia.ui.*`
FIDL libraries that do not currently specify `@available`. This will
allow the `fidl()` template to verify all libraries in the "fuchsia"
namespace in a subsequent CL.

The previous `@available` line in fuchsia.ui.annotation was added in
https://fxrev.dev/872816 but the library was in the "internal"
category at that time and should not have been marked this way.

`fuchsia.ui.focus` and `fuchsia.ui.pointer.augment` previously had
`@available(added=7)`, which was added in https://fxrev.dev/664072.
The former was in the "internal" category at that time and should not
have been marked this way. The latter was not corrected when the
category was changed in https://fxrev.dev/807610.

Add TODOs for six libraries that cannot currently be versioned because
CTF tests depend on them.

Also add missing includes to a couple tests.

Original-Bug: 364628320, 365602422
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114050
Original-Revision: 680e1bb5ebae1698d848bef29f60ee79192deb5d
GitOrigin-RevId: eb2ec94ff5fb9542ae67911503ce9ef7797eb897
Change-Id: Ie8a54516cdc1ac818106424026284a6808e2f35e