[roll] Roll fuchsia [ffx] extend ssh timeout
We are seeing ssh timeouts when connecting to emulators
(see https://b.corp.google.com/issues/364345273#comment6).
The delays are probably due to a Cavium issue (b/343550891)
in which a process (e.g. an emulator) can be delayed for
up to 15 seconds.

This change mitigates the problem, while the root cause is being
addressed.

(This change also improves an error message -- the previous version
resulted in a lot of lost time due to confusion about the error in
question.)

Original-Bug: 364345273
Original-Fixed: 366582564
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119775
Original-Revision: 8c2b383165bb16e406bee0f24f8ec26191748531
GitOrigin-RevId: 4e4278f2f3094b0b075ed19d9553c844a91a58c4
Change-Id: I187d85555b7f92a64c1eb9da1c006676063041a1