[roll] Roll scudo [scudo] Update secondary cache time-based release logic (#107507)
Secondary cache entries are now released to the OS from least recent to
most recent entries. This helps to avoid unnecessary scans of the cache
since entries ready to be released (specifically, entries that are
considered old relative to the configurable release interval) will
always be at the tail of the list of committed entries by the LRU
ordering. For this same reason, the `OldestTime` variable is no longer
needed to indicate when releases are necessary so it has been removed.

GitOrigin-RevId: d032056c68ef1305c5bc2b593f52cce7a5129107
Original-Revision: dec553452e757a89467b5aaebcd759bc0ebdb0c5
Change-Id: I1a6e9a7b59330d0c331517494c862cea48420943