[roll] Roll fuchsia [archivist] Remove double parsing of incoming structured messages
The code was previously parsing the full record to extract the timestamp
and the severity. We can just look at the header and the number after to
learn that instead of parsing the full record and then parsing it again
when reading.

There's additional clean up that seems to have been left behind regarding raw severities and verbosities. Will do in a follow-up.

Run-All-Tests: True
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119212
Original-Revision: 0bcd9ee62a2ecc7040d175524121b30d7aedba45
GitOrigin-RevId: b7c2620b0ce1292d4669b52fa9f5c15cecaf03c6
Change-Id: Ie091c39e4af558d5e8651011f6f712a14a4956dd