[roll] Roll fuchsia [windowed-stats] Don't early terminate TimeMatrix server_fut
Right now we require that the user of the library always hold onto the
TimeMatrixClient, or its corresponding server fut would be terminated.
Given that the client is only needed to register time matrices, we don't
really need the user to hold onto it forever.

In this commit, even when TimeMatrixClient is dropped, the server fut
would continue running.

Original-Bug: 350797382
Test: Added unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115584
Original-Revision: 216d37c70b480eb44ef51d068fdb1fde948bf757
GitOrigin-RevId: 94df88f39cdf1d88973b7c531504b52499708171
Change-Id: Iffbea06cc4c911c303ee093ef0af424e453320a2