[roll] Roll fuchsia [bazel_sdk] Remove @fuchsia_sdk_common mentions from shared rules.
This is the first in a series of two CLs which refactor how shared
definitions between SDK rules and the in-tree Bazel workspace
are exposed and used.

The `//build/bazel_sdk/bazel_rules_fuchsia/common` directory
contains shared definitions used by both the SDK Bazel rules and
the in-tree Bazel workspace.

When used in-tree, the directory is currently wrapped as an
external repository named `@fuchsia_sdk_common`, however this is
not required at all for OOT workspaces, which do not define it.

This CL removes a few hard-coded references to `@fuchsia_sdk_common`
for some of the functions defined in the directory, making these
usable from `@fuchsia_sdk` and `@fuchsia_clang` as well.

This is done by introducing a new `common_package_prefix` argument
to several of the functions, and adjusting only two call sites
outside of the `common` directory.

Original-Bug: 366156651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120392
Original-Revision: f5e540d0ee87be6b47c22421d554d08cd4072e5e
GitOrigin-RevId: 9d37f4dcc4a8321e21503a0fae219072d3a3f530
Change-Id: Iaf7f03c6b13b5e2a536cbfceac019bdfc342f79e