[roll] Roll fuchsia [sestarnix] Allow FsNode's in /proc/<pid> to be labeled based on a Task.
This CL turns the SecurityState of FsNode's into a container for an enum
that has 3 states:
* Uninitialized
* SecurityId
* WeakRef<Task>

This makes it possible for an FsNode to have an effective SecurityId
that tracks that of a Task, even if that Task's SecurityId changes.

This CL labels the /proc/<pid>/ directories, but not the children of
/proc/<pid>/.

Later CLs will add logic to propagate Task-based FsNode labeling to
child FsNode's as they are created.

Original-Bug: 322850026

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103653
Original-Revision: f88f181353d0bb829053bee1ec80c893925965d4
GitOrigin-RevId: f85eab42d4e602e0ab0778369514103461c348d8
Change-Id: I4eeb2d85a7438b8d776e520c8337c4821935b71a