[roll] Roll fuchsia [ffx][fho] Adding external subtool manifest support
This adds building a JSON manifest of the external subtools built as
part of the build, and a configuration key to use to pass it into
ffx. When the subtool manifest is configured, the manifest is read to
configure the external ffx subtools. If the manifest is not present,
it scans the appropriate directories looking for subtools.

By adding the manifest processing, ffx can be used hermetically in
the build, and still access external subtools.

Original-Bug: 366318546
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121714
Original-Revision: ea620acde9cb302bbcd12630c4c6c245f175bd6b
GitOrigin-RevId: 411ae5769f35438da584a26e3adee43a14753d03
Change-Id: I1fb3f206dd9b71709f92fddd9811b6dae679b37e