[roll] Roll fuchsia [usb][dwc3] move uep lock inside of the endpoint
There is a 1:1 relationship between uep and underlying endpoint struct,
so this can be moved behind the endpoint struct. This is a necessary
step to add the Endpoint protocol impl, and it got messy doing it at
the same time. I refactored this change out for simplicity.

Test: Boot kola and ssh login over CDC
Original-Bug: b/350802129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121455
Original-Revision: d5138bdd5eeacb37e17debb0c66ffceb4fa64c50
GitOrigin-RevId: ee768e6e1b5d6b06248fd399c99471089b4fb5ff
Change-Id: I723e6aa2e2d232335fc3c9656624db25de04b81d