[roll] Roll fuchsia [uart][geni] Fix last byte handling of the RX FIFO.
Previously, the zircon Qualcomm GENI UART driver checked the bit 31
of the RX FIFO status register to determine if the last word contains
bits that are fewer than the hardware word width; if true, it read
the bit 30-28 to get the number of valid bytes in the last word.
If the bit 31 was true and the number of valid bytes in the last word
was zero, the driver would not read the last word from the FIFO.

The register definition above is incorrect. Experiments show that:

The bit 31 of the RX FIFO status register indicates whether the RX
FIFO times out.
The bits 30-28 of the RX FIFO status register are the number of valid
bytes in the last word, only if the RX times out and the bits 30-28
are not zero.

The previous uart driver did not read the last word from the FIFO.
The hardware wouldn't enqueue new bytes into the FIFO, therefore the
RX FIFO got stuck forever.

Previous changes (I56cc1ff8124cebb38e3e2aa897dc7a6b7c23e15f and
I7321fff6f0749948f80b78f319cf338fcfd5d4fc) partially fixed this issue
by changing the watermark interrupt trigger patterns but it didn't
fully solve this problem.

This change:
(1) Moves the FIFO clear operations to the end of the interrupt
    handler. This makes sure that the status registers won't be
    changed in the middle of the interrupt handler.

(2) Fixes the FifoStatusRegister definition.

(3) Read partial bytes from the FIFO only if
    valid_bytes_in_last_word_if_not_zero is not zero.

Test: uart-unittests
Test: We test the serial terminal behavior comprehensively on the
      XXXXXX board. Steps:

      1. Run the following script to generate input of length 4..400.
      for len in `seq 4 400`; do
        chars=$(
          for i in {1..$(($len - 3))}; do echo -n =; done
          printf "%03d" $len
        )
        echo $chars
      done

      2. Paste each line to the serial terminal.

      3. Make sure that the serial terminal input is not stuck. It can
	 still read new characters.

Original-Fixed: 361345380
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110528
Original-Revision: bf0edce246955747a2fbfd7df0f92a57948430dd
GitOrigin-RevId: 77af8ea67232a50483af0d9c3d08b3b33612f1bf
Change-Id: Id1d24c956f9e0c28542d409d6c363cc9ceb70fd7