[roll] Roll fuchsia [zx][rs] Add API for boot time and timers.
Note that these APIs do not yet read from CLOCK_BOOT because it is
not yet stable. This is fine as long as the monotonic clock does
not yet pause during suspends, as the two timelines are currently
equivalent. The naming for these APIs is not yet final and may be
revised before the relevant RFCs are ratified.

This will allow clients to begin migrating to the new APIs in
advance of a future change which will switch these APIs to be
based on the new timeline.

Original-Bug: 328306129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106092
Original-Revision: 1a59a7ed602ff89591fde43f6eb77ca1bac615d1
GitOrigin-RevId: bd63512346cc653b47f568456d455748b5f17ef0
Change-Id: I299e5c1713e9f8d7a0cccab498286bc296e2a0be