[roll] Roll fuchsia [pkg] Deduplicating RepoStorageType, ConflictMode
This removes the recently added pkg:: enums that mirror
the FIDL enums for the StorageType and Alias conflict mode. These
mirrored versions are used to add serialization to the enum.

There is still a mirror in fidl_ffx_developer_ext, but that can
be handled in a follow-up CL.

Original-Bug: 367668384
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120573
Original-Revision: 19815e7eb7b9c9ef160cac16a2c9bd698398eeb0
GitOrigin-RevId: 5bbe3ac8b689a81a88477b0ef2a23edb566d5c36
Change-Id: Ia9e1a9ff8e8273dc74f20b8dd2d5e0eefaf5a19e