[roll] Roll fuchsia [fxfs] move has_overwrite_extents flag to attribute value
Instead of having it as part of the main object value, move the flag to
be per-attribute. This is primarily an optimization - when opening an
object we already look up the data attribute, so we can use that
existing check to find this flag as well.

Original-Bug: 293943124
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122293
Original-Revision: 0ca1104332b92f2a21662d571cc2ce1696f6887d
GitOrigin-RevId: 33e9db5f17615568b6768fa49c1475b2cfd8da91
Change-Id: I7a335c18829f826c0db346ba33b334838d787fe3